Adjust types in searchcore and document module.

+++ searchcore/src/vespa/searchcore/proton/docsummary/summarycompacttarget.cpp
@@ -58,7 +58,7 @@ IFlushTarget::DiskGain
 SummaryCompactTarget::getApproxDiskGain() const
 {
     uint64_t total(_docStore.getDiskFootprint());
-    return DiskGain(total, total - std::min(total, _docStore.getMaxCompactGain()));
+    return DiskGain(total, total - std::min(total, static_cast<uint64_t>(_docStore.getMaxCompactGain())));
 }
 
 IFlushTarget::Time
